requirements.txt
flask==2.3.3
Flask-Cors
flask-cors==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas
python-dotenv==1.0.0
requests==2.31.0
